+++ tests/test_doccmd.py
@@ -1895,3 +1895,121 @@ def test_de_duplication_source_files_and_dirs(tmp_path: Path) -> None:
 
     assert result.stdout == expected_output
     assert result.stderr == ""
+
+
+def test_max_depth(tmp_path: Path) -> None:
+    """
+    The --max-depth option limits the depth of directories to search for files.
+    """
+    runner = CliRunner(mix_stderr=False)
+    rst_file = tmp_path / "example.rst"
+    rst_content = """\
+    .. code-block:: python
+
+        rst_1_block
+    """
+    rst_file.write_text(data=rst_content, encoding="utf-8")
+
+    sub_directory = tmp_path / "subdir"
+    sub_directory.mkdir()
+    rst_file_in_sub_directory = sub_directory / "subdir_example.rst"
+    subdir_rst_content = """\
+    .. code-block:: python
+
+        rst_subdir_1_block
+    """
+    rst_file_in_sub_directory.write_text(
+        data=subdir_rst_content,
+        encoding="utf-8",
+    )
+
+    sub_sub_directory = sub_directory / "subsubdir"
+    sub_sub_directory.mkdir()
+    rst_file_in_sub_sub_directory = sub_sub_directory / "subsubdir_example.rst"
+    subsubdir_rst_content = """\
+    .. code-block:: python
+
+        rst_subsubdir_1_block
+    """
+    rst_file_in_sub_sub_directory.write_text(
+        data=subsubdir_rst_content,
+        encoding="utf-8",
+    )
+
+    arguments = [
+        "--language",
+        "python",
+        "--no-pad-file",
+        "--command",
+        "cat",
+        "--max-depth",
+        "1",
+        str(tmp_path),
+    ]
+    result = runner.invoke(
+        cli=main,
+        args=arguments,
+        catch_exceptions=False,
+    )
+    assert result.exit_code == 0, result.stderr
+    expected_output = textwrap.dedent(
+        text="""\
+        rst_1_block
+        """,
+    )
+
+    assert result.stdout == expected_output
+    assert result.stderr == ""
+
+    arguments = [
+        "--language",
+        "python",
+        "--no-pad-file",
+        "--command",
+        "cat",
+        "--max-depth",
+        "2",
+        str(tmp_path),
+    ]
+    result = runner.invoke(
+        cli=main,
+        args=arguments,
+        catch_exceptions=False,
+    )
+    assert result.exit_code == 0, result.stderr
+    expected_output = textwrap.dedent(
+        text="""\
+        rst_1_block
+        rst_subdir_1_block
+        """,
+    )
+
+    assert result.stdout == expected_output
+    assert result.stderr == ""
+
+    arguments = [
+        "--language",
+        "python",
+        "--no-pad-file",
+        "--command",
+        "cat",
+        "--max-depth",
+        "3",
+        str(tmp_path),
+    ]
+    result = runner.invoke(
+        cli=main,
+        args=arguments,
+        catch_exceptions=False,
+    )
+    assert result.exit_code == 0, result.stderr
+    expected_output = textwrap.dedent(
+        text="""\
+        rst_1_block
+        rst_subdir_1_block
+        rst_subsubdir_1_block
+        """,
+    )
+
+    assert result.stdout == expected_output
+    assert result.stderr == ""