+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Next
 ----
 
+* Add a ``--max-depth`` option for recursing into directories.
+
 2024.11.06
 ----------